+++ apps/frontend/src/app/3_organisms/FastBtcDialog/components/BoltzSendFlow/components/StatusScreen.tsx
@@ -26,7 +26,7 @@ import {
   Status,
   StatusEnum,
   SubmarineSwapResponse,
-} from '../../../utils/boltz';
+} from '../../../utils/boltz/boltz.types';
 import { BoltzStatus } from './BoltzStatus';
 import { getDescription, getTitle } from './StatusScreen.utils';
