+++ apps/frontend/src/app/3_organisms/FastBtcDialog/contexts/deposit-boltz-context.ts
@@ -1,6 +1,6 @@
 import { createContext, Dispatch, SetStateAction } from 'react';
 
-import { ReverseSwapPair } from '../utils/boltz';
+import { ReverseSwapPair } from '../utils/boltz/boltz.types';
 
 export enum DepositBoltzStep {
   MAIN,
