+++ apps/frontend/src/app/3_organisms/FastBtcDialog/hooks/useWithdrawBoltzConfig.ts
@@ -1,18 +1,17 @@
 import { useCallback, useEffect, useMemo, useState } from 'react';
 
-import { getContracts } from '../../Boltz/Boltz.utils';
 import {
   defaultValue,
   WithdrawBoltzContextStateType,
 } from '../contexts/withdraw-boltz-context';
-import { boltz } from '../utils/boltz';
+import { boltz } from '../utils/boltz/boltz.client';
 
 export function useWithdrawBoltzConfig() {
   const [state, setState] =
     useState<WithdrawBoltzContextStateType>(defaultValue);
 
   const getBoltzLimits = useCallback(async () => {
-    const contracts = await getContracts();
+    const contracts = await boltz.getContracts();
     const pair = await boltz
       .getSubmarineSwapPairs()
       .then(pairs => pairs['RBTC']['BTC']);
