+++ apps/frontend/src/utils/helpers.ts
@@ -177,3 +177,6 @@ export const areValuesIdentical = (
 
   return Math.abs(firstValue.sub(secondValue).toNumber()) < epsilon;
 };
+
+export const prefix0x = (val: string) =>
+  val.startsWith('0x') ? val : `0x${val}`;
